requirements.txt
aiohttp==3.7.4.post0
aiomysql==0.0.21
aiosqlite==0.16.1
anyio==3.1.0
asgiref==3.3.4
async-timeout==3.0.1
asynctest==0.13.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
fastapi-pagination==0.7.2
h11==0.12.0
httpcore==0.13.5
httpx==0.18.1
idna==2.10
iniconfig==1.1.1
iso8601==0.1.14
multidict==5.1.0
mysqlclient==2.0.3
packaging==20.9
pluggy==0.13.1
py==1.10.0
pydantic==1.8.2
PyMySQL==0.9.3
pyparsing==2.4.7
pypika-tortoise==0.1.1
pytest==6.2.4
pytest-asyncio==0.15.1
python-dotenv==0.17.1
pytz==2020.5
requests==2.25.1
rfc3986==1.5.0
sniffio==1.2.0
starlette==0.13.6
toml==0.10.2
tortoise-orm==0.17.2
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.13.4
yarl==1.6.3
